Move logic from authentication controller
