+++ app/services/json_web_token.rb
