+++ config/routes.rb
@@ -1,7 +1,6 @@
 Rails.application.routes.draw do
   resources :users, param: :_username
-  post '/auth/login', to: 'authentication#login'
-
+  resources :authentications, only: [:create]
   resources :classrooms
   resources :students, only: [:index, :show]
   resources :presentation_areas, only: [:index, :show]
